requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
sqlalchemy>=2.0.30
google-generativeai>=0.7.0

